# sources/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация по API",
        "description": "API документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/authorization/": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация профиля: выдаёт пользователя и Bearer токен",
                "description": "Возвращает информацию о профиле и Bearer токен",
                "operationId": "53f22ac8c86268a3b36ebf1c7fda9c9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+7 (999) 999-99-99"
                                    },
                                    "pincode": {
                                        "type": "string",
                                        "example": "последние 6 цифр звонилки"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "УСПЕХ!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "attributes": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Authorization success"
                                                },
                                                "user": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "properties": {
                                                        "type": {
                                                            "type": "string",
                                                            "example": "Bearer"
                                                        },
                                                        "access_token": {
                                                            "type": "string",
                                                            "example": "1|eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
...
        "/api/v1/candidates/get-statuses": {
            "get": {
                "tags": [
                    "CandidateProfiles"
                ],
                "summary": "Получение статусов для анкеты кандидата",
                "description": "Возвращение JSON объекта",
                "operationId": "9f475a43786b93e17a52ac8f68ab30c3",
                "responses": {
                    "200": {
                        "description": "УСПЕХ!"
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Анкета не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/crm/client-transfer": {
            "post": {
                "tags": [
                    "Crm"
                ],
                "summary": "Отправка заявки на передачу клиента в ЦРМ",
                "description": "Возвращение JSON объекта",
                "operationId": "63e8c3c2659bbf4e21273fc923c002b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "last_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "client_last_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "client_first_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "client_middle_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "client_phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "УСПЕХ!"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}